Allow 64-bit sizes to be passed to robag max_size (#865)

+++ tools/rosbag/include/rosbag/recorder.h
@@ -103,7 +103,7 @@ struct ROSBAG_DECL RecorderOptions
     uint32_t        chunk_size;
     uint32_t        limit;
     bool            split;
-    uint32_t        max_size;
+    uint64_t        max_size;
     ros::Duration   max_duration;
     std::string     node;
     unsigned long long min_space;
