+++ tools/rosbag/src/record.cpp
@@ -63,7 +63,7 @@ rosbag::RecorderOptions parseOptions(int argc, char** argv) {
       ("lz4", "use LZ4 compression")
       ("split", po::value<int>()->implicit_value(0), "Split the bag file and continue recording when maximum size or maximum duration reached.")
       ("topic", po::value< std::vector<std::string> >(), "topic to record")
-      ("size", po::value<int>(), "The maximum size of the bag to record in MB.")
+      ("size", po::value<uint64_t>(), "The maximum size of the bag to record in MB.")
       ("duration", po::value<std::string>(), "Record a bag of maximum duration in seconds, unless 'm', or 'h' is appended.")
       ("node", po::value<std::string>(), "Record all topics subscribed to by a specific node.");
 
@@ -217,7 +217,7 @@ rosbag::RecorderOptions parseOptions(int argc, char** argv) {
     }
     if (vm.count("size"))
     {
-      opts.max_size = vm["size"].as<int>() * 1048576;
+      opts.max_size = vm["size"].as<uint64_t>() * 1048576;
       if (opts.max_size <= 0)
         throw ros::Exception("Split size must be 0 or positive");
     }
